Moved SVG parsing to extension. Fixed bug with parsing SVG polylines
