+++ CMakeLists.txt
@@ -42,6 +42,7 @@ set(SOURCES ${SOURCES} Source/Graphic/HCColor.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCRaster.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCContour.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCPath.c)
+set(SOURCES ${SOURCES} Source/Graphic/HCPath+SVG.c)
 set(SOURCES ${SOURCES} Source/JSON/HCJSON.c)
 set(SOURCES ${SOURCES} Source/Thread/HCThread.c)
 set(SOURCES ${SOURCES} Source/Thread/HCLock.c)
