+++ Source/Graphic/HCPath_Internal.h
@@ -25,7 +25,7 @@ typedef struct HCPath {
 //----------------------------------------------------------------------------------------------------------------------------------
 // MARK: - Construction
 //----------------------------------------------------------------------------------------------------------------------------------
-void HCPathInit(void* memory, const char* path);
+void HCPathInit(void* memory);
 void HCPathDestroy(HCPathRef self);
 
 //----------------------------------------------------------------------------------------------------------------------------------
@@ -45,10 +45,5 @@ void HCPathIntersects(void* context, HCBoolean* stopSearching, HCPathRef path, H
 //----------------------------------------------------------------------------------------------------------------------------------
 HCBoolean HCPathFilterOpen(void* context, HCPathRef path);
 HCBoolean HCPathFilterClosed(void* context, HCPathRef path);
-
-//----------------------------------------------------------------------------------------------------------------------------------
-// MARK: - Path Parsing
-//----------------------------------------------------------------------------------------------------------------------------------
-void HCPathParse(HCPathRef self, const char* path);
 
 #endif /* HCPath_Internal_h */
