+++ Source/Graphic/HCRaster.h
@@ -12,6 +12,7 @@
 #include "../Core/HCObject.h"
 #include "HCColor.h"
 #include "HCPath.h"
+#include "HCPath+SVG.h"
 
 //----------------------------------------------------------------------------------------------------------------------------------
 // MARK: - Object Type
